chore: extract hook and update og image
