+++ apps/web/src/app/api/og/route.tsx
@@ -8,7 +8,7 @@ const size = {
 };
 
 const TITLE = "Open Status";
-const DESCRIPTION = "Open Source Alternative to ...";
+const DESCRIPTION = "An Open Source Alternative for your next Status Page";
 
 const interRegular = fetch(
   new URL("../../../public/fonts/Inter-Regular.ttf", import.meta.url)
@@ -30,8 +30,18 @@ export async function GET(req: Request) {
 
   return new ImageResponse(
     (
-      <div tw="flex flex-col bg-white items-center justify-center w-full h-full">
-        <div tw="flex flex-col items-center">
+      <div tw="relative flex flex-col bg-white items-center justify-center w-full h-full">
+        <div
+          tw="flex w-full h-full absolute inset-0"
+          // not every css variable is supported
+          style={{
+            backgroundImage: "radial-gradient(#94a3b8 10%, transparent 10%)",
+            backgroundPosition: "0px 0px, 6px 6px",
+            backgroundSize: "12px 12px",
+            filter: "blur(1px)", // to be discussed... couldn't put it inside the content container
+          }}
+        ></div>
+        <div tw="max-w-2xl relative flex flex-col items-center rounded-lg border border-slate-200 p-6 overflow-hidden bg-white bg-opacity-80">
           <h1 style={{ fontFamily: "Cal" }} tw="text-6xl text-center">
             {title}
           </h1>
